requirement.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
djangorestframework==3.16.0
docutils==0.21.2
filetype==1.2.0
idna==3.10
Kivy==2.3.1
Kivy-Garden==0.1.5
kivymd==1.2.0
pillow==11.2.1
Pygments==2.19.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0
